cast on check

+++ src/transformers/models/bart/modeling_tf_bart.py
@@ -760,10 +760,10 @@ class TFBartEncoder(tf.keras.layers.Layer):
             with context_manager:
                 if tf.executing_eagerly():
                     # Note: tf.gather, on which the embedding layer is based, won't check positive out of bound
-                    # indices on GPU, returning zeros instead. This is a dangeroud silent behavior.
+                    # indices on GPU, returning zeros instead. This is a dangerous silent behavior.
                     tf.debugging.assert_less(
                         input_ids,
-                        self.embed_tokens.input_dim,
+                        tf.cast(self.embed_tokens.input_dim, dtype=input_ids.dtype),
                         message=(
                             "input_ids must be smaller than the embedding layer's input dimension (got"
                             f" {tf.math.reduce_max(input_ids)} >={self.embed_tokens.input_dim})"
@@ -967,10 +967,10 @@ class TFBartDecoder(tf.keras.layers.Layer):
             with context_manager:
                 if tf.executing_eagerly():
                     # Note: tf.gather, on which the embedding layer is based, won't check positive out of bound
-                    # indices on GPU, returning zeros instead. This is a dangeroud silent behavior.
+                    # indices on GPU, returning zeros instead. This is a dangerous silent behavior.
                     tf.debugging.assert_less(
                         input_ids,
-                        self.embed_tokens.input_dim,
+                        tf.cast(self.embed_tokens.input_dim, dtype=input_ids.dtype),
                         message=(
                             "input_ids must be smaller than the embedding layer's input dimension (got"
                             f" {tf.math.reduce_max(input_ids)} >={self.embed_tokens.input_dim})"
